requirements.txt
requests>=2.31.0
urllib3<2
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2
yt-dlp>=2025.2.19
praw>=7.8.1
